WIP: Wrap API consumption into several services
