+++ src/pages/Category.js
@@ -5,12 +5,9 @@ import Navbar from '../components/Navbar.js'
 import SideBarMenu from '../components/SideBarMenu.js'
 import CategoryTable from '../components/Category/CategoryTable.js'
 import CategoryForm from '../components/Category/CategoryForm.js'
-import { selectUserAuthToken } from '../store/selectors/user.js'
-import {
-  createShowSuccessNotificationAction,
-  createShowErrorNotificationAction
-} from '../store/actions/notification.js'
-import api from '../api'
+import CategoryService from '../services/Category/CategoryService'
+
+const categoryService = new CategoryService()
 
 const Category = () => {
 
@@ -19,15 +16,9 @@ const Category = () => {
   const { state, dispatch } = useContext(AppContext)
   const history = useHistory()
 
-  const headers = {
-    'Auth': selectUserAuthToken(state),
-    'Content-Type': 'application/json'
-  }
-
 
   const fetchCategories = async () => {
-    const categoriesResponse = await api.getCategories(headers)
-    categoriesResponse && !categoriesResponse.error ? setCategories(categoriesResponse) : history.push('/login')
+    await categoryService.fetchCategories(setCategories, state, history)
   }
 
 
@@ -37,64 +28,17 @@ const Category = () => {
 
 
   const onCreateCategory = async (categoryData) => {
-
-    const onSuccess = response => {
-      dispatch(createShowSuccessNotificationAction({
-        header: '¡Categoría Creada!',
-        message: 'La categoría se creó con éxito'
-      }))
-      categories.push(response)
-      setCategories(categories)
-    }
-
-    const onError = error => {
-      dispatch(createShowErrorNotificationAction({
-        header: '¡Error!',
-        message: 'Ha ocurrido un error cuando se intentaba crear una categoría'
-      }))
-    }
-
-    await api.createCatetory(categoryData, headers, onSuccess, onError)
+    await categoryService.onCreateCategory(categoryData, dispatch, categories, setCategories, state)
   }
 
 
   const onDeleteCategory = async (idx, categoryData) => {
-
-    const onSuccess = response => {
-      dispatch(createShowSuccessNotificationAction({
-        header: '¡Categoría Eliminada!',
-        message: 'La categoría se elimino con éxito'
-      }))
-    }
-
-    const onError = error => {
-      dispatch(createShowErrorNotificationAction({
-        header: '¡Error!',
-        message: 'Ha ocurrido un error cuando se intentaba eliminar una categoría'
-      }))
-    }
-
-    await api.removeCategory(categoryData.id, headers, onSuccess, onError)
-    categories.splice(idx, 1)
+    await categoryService.onDeleteCategory(idx, categoryData, categories, dispatch, state)
   }
 
 
   const onEditCategory = async (category) => {
-    const onSuccess = response => {
-      dispatch(createShowSuccessNotificationAction({
-        header: '¡Categoría Modificada!',
-        message: 'La categoría se modificó con éxito'
-      }))
-    }
-
-    const onError = error => {
-      dispatch(createShowErrorNotificationAction({
-        header: '¡Error!',
-        message: 'Ha ocurrido un error cuando se intentaba modificar una categoría'
-      }))
-    }
-
-    await api.updateCategory(category.id, category,  headers, onSuccess, onError)
+    await categoryService.onEditCategory(category, dispatch, state)
   }
